cambios en tareas
- agregar validador de tags en la vista principal de las tareas
- agregar lógica de etiquetas en modal de filtros de tareas
- revisar lógica de tiempo transcurrido de la tarea
