+++ src/app/[locale]/(home)/tools/(tool)/tasks/components/filters/AddFields.jsx
@@ -1,20 +1,20 @@
 'use client';
 import React, { Fragment } from 'react';
-import { Menu, Transition } from '@headlessui/react';
+import { Menu, Transition, MenuButton, MenuItems } from '@headlessui/react';
 import { PlusIcon } from '@heroicons/react/20/solid';
 import { useTranslation } from 'react-i18next';
 import NewFields from './NewFields';
 
-const AddFields = ({append, remove, fields}) => {
+const AddFields = ({ append, remove, fields }) => {
 	const { t } = useTranslation();
-	
+
 	return (
 		<Menu as="div" className="relative inline-block">
 			<div>
-				<Menu.Button className="inline-flex text-gray-60 bg-transparent text-xs font-semibold gap-2 mt-1.5">
+				<MenuButton className="inline-flex text-gray-60 bg-transparent text-xs font-semibold gap-2 mt-1.5">
 					<PlusIcon className="h-4 w-4" />
 					{t('contacts:filters:add-field')}
-				</Menu.Button>
+				</MenuButton>
 			</div>
 			<Transition
 				as={Fragment}
@@ -25,11 +25,11 @@ const AddFields = ({append, remove, fields}) => {
 				leaveFrom="transform opacity-100 scale-100"
 				leaveTo="transform opacity-0 scale-95"
 			>
-				<Menu.Items
+				<MenuItems
 					className={`absolute left-0 mt-2 rounded-md bg-blue-50 shadow-lg ring-1 ring-black/5 focus:outline-none z-50 w-96`}
 				>
-                    <NewFields append={append} remove={remove} fields={fields}/>
-				</Menu.Items>
+					<NewFields append={append} remove={remove} fields={fields} />
+				</MenuItems>
 			</Transition>
 		</Menu>
 	);
