+++ src/components/notifier/drawer/NotifyDrawer.jsx
@@ -37,7 +37,7 @@ export default function NotifyDrawer() {
               <div className="flex h-full divide-y divide-slate-400 flex-col overflow-y-scroll bg-zinc-300 py-6 shadow-xl">
                 <div className="px-4 sm:px-6">
                   <DialogTitle className="text-base font-semibold leading-6 text-gray-900 flex gap-2 items-center ">
-                    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" fill="currentColor" class="size-5 w-8 h-8 text-slate-800">
+                    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" fill="currentColor" className="size-5 w-8 h-8 text-slate-800">
                       <path fill-rule="evenodd" d="M10 2a6 6 0 0 0-6 6c0 1.887-.454 3.665-1.257 5.234a.75.75 0 0 0 .515 1.076 32.91 32.91 0 0 0 3.256.508 3.5 3.5 0 0 0 6.972 0 32.903 32.903 0 0 0 3.256-.508.75.75 0 0 0 .515-1.076A11.448 11.448 0 0 1 16 8a6 6 0 0 0-6-6ZM8.05 14.943a33.54 33.54 0 0 0 3.9 0 2 2 0 0 1-3.9 0Z" clip-rule="evenodd" />
                     </svg>
                     <div className='text-lg'>Notificaciones</div>
